requirements.txt


bcrypt==3.1.5
certifi==2018.11.29
cffi==1.11.5
Django==1.11.23
django-registration-redux==1.4
names==0.3.0
olefile==0.46
pbr==5.4.2
Pillow==5.3.0
pycparser==2.19
pytz==2018.7
selenium==3.141.0
six==1.12.0
stevedore==1.30.1
urllib3==1.24.1
virtualenv==16.7.2
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.5
wincertstore==0.2
